--- packages/buidler/.eslintrc.js
@@ -1,23 +1,0 @@
-module.exports = {
-  env: {
-    mocha: true,
-  },
-  extends: ["airbnb", "plugin:prettier/recommended"],
-  plugins: ["babel"],
-  rules: {
-    "prettier/prettier": ["error"],
-    "import/extensions": [
-      "error",
-      "ignorePackages",
-      {
-        js: "never",
-        ts: "never",
-      },
-    ],
-    "import/prefer-default-export": "off",
-    "prefer-destructuring": "off",
-    "prefer-template": "off",
-    "no-console": "off",
-    "func-names": "off",
-  },
-};
