--- packages/buidler/package.json
@@ -1,34 +1,0 @@
-{
-  "name": "@scaffold-eth/buidler",
-  "version": "1.0.0",
-  "main": "index.js",
-  "license": "MIT",
-  "devDependencies": {
-    "eslint": "^7.5.0",
-    "eslint-config-airbnb": "^18.2.0",
-    "eslint-config-prettier": "^6.11.0",
-    "eslint-plugin-babel": "^5.3.1",
-    "eslint-plugin-prettier": "^3.1.4"
-  },
-  "dependencies": {
-    "@nomiclabs/buidler": "^1.4.1",
-    "@nomiclabs/buidler-ethers": "^2.0.0",
-    "@nomiclabs/buidler-waffle": "^2.0.0",
-    "@openzeppelin/contracts": "^3.1.0",
-    "chai": "^4.2.0",
-    "chalk": "^4.1.0",
-    "ethereum-waffle": "^3.0.2",
-    "ethers": "^5.0.7",
-    "node-watch": "^0.6.4"
-  },
-  "scripts": {
-    "chain": "buidler node",
-    "test": "buidler test --network buidlerevm",
-    "compile": "buidler compile",
-    "deploy": "buidler run scripts/deploy.js && buidler run scripts/publish.js",
-    "watch": "node scripts/watch.js",
-    "accounts": "buidler accounts",
-    "balance": "buidler balance",
-    "send": "buidler send"
-  }
-}
