+++ packages/contracts/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@scaffold-eth/contracts",
+  "name": "@signchain/contracts",
   "version": "1.0.0",
   "main": "./src/index.js"
 }
