+++ packages/hardhat/artifacts/DocumentRegistry.address
@@ -1,0 +1,1 @@
+0x135EEfE996D85B0567200739440D6C52ba47a015
