+++ packages/hardhat/artifacts/DocumentRegistry.args
@@ -1,0 +1,1 @@
+0x
