+++ packages/hardhat/artifacts/Signchain.address
@@ -1,0 +1,1 @@
+0xDC06B41Bddb8A22Ef69e329D9954222032583226
