+++ packages/hardhat/artifacts/Signchain.args
@@ -1,0 +1,1 @@
+0x
