+++ packages/hardhat/artifacts/SigningModule.address
@@ -1,0 +1,1 @@
+0x9b7D68289807a9C0Cac14868d3cA1EB27aB7b0Bd
