+++ packages/hardhat/artifacts/SigningModule.args
@@ -1,0 +1,1 @@
+0x
