+++ packages/hardhat/artifacts/contracts/Signchain.sol/Signchain.dbg.json
@@ -1,0 +1,4 @@
+{
+  "_format": "hh-sol-dbg-1",
+  "buildInfo": "../../build-info/18d683737bde0051eed2c2d1b000f9e3.json"
+}
