+++ packages/hardhat/artifacts/hardhat/console.sol/console.json
@@ -1,0 +1,10 @@
+{
+  "_format": "hh-sol-artifact-1",
+  "contractName": "console",
+  "sourceName": "hardhat/console.sol",
+  "abi": [],
+  "bytecode": "0x60566023600b82828239805160001a607314601657fe5b30600052607381538281f3fe73000000000000000000000000000000000000000030146080604052600080fdfea26469706673582212204d14aaaec70c060c110fb6a706ce63b40e0bf5ca795042848ada68b1d9283cb764736f6c63430006070033",
+  "deployedBytecode": "0x73000000000000000000000000000000000000000030146080604052600080fdfea26469706673582212204d14aaaec70c060c110fb6a706ce63b40e0bf5ca795042848ada68b1d9283cb764736f6c63430006070033",
+  "linkReferences": {},
+  "deployedLinkReferences": {}
+}
