+++ packages/hardhat/cache/last-solc-config.json
@@ -1,0 +1,10 @@
+{
+  "solc": {
+    "version": "0.6.6",
+    "optimizer": {
+      "enabled": true,
+      "runs": 200
+    }
+  },
+  "buidlerVersion": "1.4.1"
+}
